used API, Arranged the folder structure
