+++ package.json
@@ -12,10 +12,12 @@
   "dependencies": {
     "@ant-design/icons": "^5.6.0",
     "antd": "^5.23.3",
+    "axios": "^1.7.9",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.4.0",
-    "react-router": "^7.1.5"
+    "react-router": "^7.1.5",
+    "react-router-dom": "^7.1.5"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
