+++ src/Component/Responsive/ResponsiveContent/Content.css
@@ -1,0 +1,83 @@
+
+@media (max-width: 1200px) {
+    .slider-container {
+      padding: 30px 15px !important;
+    }
+  
+    .slider-cards {
+      justify-content: center !important;
+    }
+  
+    .slider-card {
+      width: 20% !important;
+    }
+  }
+  
+  /* Tablet Devices (max-width: 1024px) */
+  @media (max-width: 1024px) {
+    .slider-container {
+      padding: 25px 10px !important;
+    }
+  
+    .slider-header {
+      flex-direction: column !important;
+      align-items: center !important;
+      text-align: center !important;
+    }
+  
+    .slider-header h1 {
+      font-size: 1.8em !important;
+    }
+  
+    .slider-controls {
+      margin-top: 10px !important;
+    }
+  
+    .slider-cards {
+      flex-wrap: wrap !important;
+      justify-content: center !important;
+    }
+  
+    .slider-card {
+      width: 30% !important;
+      max-width: 200px !important;
+    }
+  
+    .slider-card img {
+      width: 140px !important;
+      height: 140px !important;
+    }
+  }
+  
+  /* Mobile Devices (max-width: 600px) */
+  @media (max-width: 600px) {
+    .slider-container {
+      padding: 20px 5px !important;
+    }
+  
+    .slider-header h1 {
+      font-size: 1.5em !important;
+    }
+  
+    .slider-controls {
+      display: flex !important;
+      flex-direction: column !important;
+      align-items: center !important;
+    }
+  
+    .slider-cards {
+      flex-wrap: wrap !important;
+      justify-content: center !important;
+    }
+  
+    .slider-card {
+      width: 90% !important;
+      max-width: 180px !important;
+    }
+  
+    .slider-card img {
+      width: 120px !important;
+      height: 120px !important;
+    }
+  }
+
