+++ src/Component/Responsive/ResponsiveFooter/Footer.css
@@ -1,0 +1,42 @@
+/* Tablets and smaller screens (1024px and below) */
+@media screen and (max-width: 1024px) {
+    .footer-container {
+      flex-wrap: wrap;
+      gap: 30px;
+      justify-content: space-around;
+    }
+  }
+  
+  /* Mobile screens (768px and below) */
+  @media screen and (max-width: 768px) {
+    .footer-container {
+      flex-direction: column;
+      align-items: center;
+      text-align: center;
+    }
+  
+    .footer-social-icons {
+      justify-content: center;
+    }
+  }
+  
+  /* Small mobile screens (480px and below) */
+  @media screen and (max-width: 480px) {
+    .footer-container {
+      gap: 20px;
+      padding: 10px;
+    }
+  
+    .footer-title {
+      font-size: 1.2em;
+    }
+  
+    .footer-list a {
+      font-size: 1em;
+    }
+  
+    .footer-subscribe p {
+      font-size: 1em;
+    }
+  }
+
