+++ src/Pages/Authentication/Login.jsx
@@ -1,0 +1,43 @@
+import { Button, Form, Input } from "antd";
+const onFinish = (values) => {
+  console.log("Success:", values);
+};
+const onFinishFailed = (errorInfo) => {
+  console.log("Failed:", errorInfo);
+};
+const Login = () => (
+  <Form onFinish={onFinish} onFinishFailed={onFinishFailed} autoComplete="off">
+    <Form.Item
+      label="Username"
+      name="username"
+      rules={[
+        {
+          required: true,
+          message: "Please input your username!",
+        },
+      ]}
+    >
+      <Input />
+    </Form.Item>
+
+    <Form.Item
+      label="Password"
+      name="password"
+      rules={[
+        {
+          required: true,
+          message: "Please input your password!",
+        },
+      ]}
+    >
+      <Input.Password />
+    </Form.Item>
+    <Form.Item label={null}>
+      <Button type="primary" htmlType="submit" style={{ float: "right" }}>
+        Submit
+      </Button>
+    </Form.Item>
+  </Form>
+);
+
+export default Login;
