+++ src/assets/Images/ImageComponent/Images.jsx
@@ -1,0 +1,76 @@
+import OrganicLogo from "../../Images/logo.svg";
+import Banner from "../../Images/banner-1.jpg";
+import CategoryImage1 from "../../Images/category-thumb-1.jpg";
+import CategoryImage2 from "../../Images/category-thumb-2.jpg";
+import CategoryImage3 from "../../Images/category-thumb-3.jpg";
+import CategoryImage4 from "../../Images/category-thumb-4.jpg";
+import CategoryImage5 from "../../Images/category-thumb-5.jpg";
+import CategoryImage6 from "../../Images/category-thumb-6.jpg";
+import CategoryImage7 from "../../Images/category-thumb-7.jpg";
+import CategoryImage8 from "../../Images/category-thumb-8.jpg";
+import ProductThumb1 from "../../Images/product-thumb-1.png";
+import ProductThumb2 from "../../Images/product-thumb-2.png";
+import ProductThumb3 from "../../Images/product-thumb-3.png";
+import ProductThumb4 from "../../Images/product-thumb-4.png";
+import ProductThumb5 from "../../Images/product-thumb-5.png";
+import ProductThumb6 from "../../Images/product-thumb-6.png";
+import ProductThumb7 from "../../Images/product-thumb-7.png";
+import ProductThumb8 from "../../Images/product-thumb-8.png";
+import ProductThumb9 from "../../Images/product-thumb-9.png";
+import ProductThumb10 from "../../Images/product-thumb-10.png";
+import ProductThumb11 from "../../Images/product-thumb-11.png";
+import ProductThumb12 from "../../Images/product-thumb-12.png";
+import ProductThumb13 from "../../Images/product-thumb-13.png";
+import ProductThumb14 from "../../Images/product-thumb-14.png";
+import ProductThumb15 from "../../Images/product-thumb-15.png";
+import ProductThumb16 from "../../Images/product-thumb-16.png";
+import ProductThumb17 from "../../Images/product-thumb-17.png";
+import ProductThumb18 from "../../Images/product-thumb-18.png";
+import ProductThumb19 from "../../Images/product-thumb-19.png";
+import ProductThumb20 from "../../Images/product-thumb-20.png";
+import ProductThumb21 from "../../Images/product-thumb-21.png";
+import ProductThumb22 from "../../Images/product-thumb-22.png";
+import BgDiscountForm from "../../Images/banner-newsletter.jpg";
+import BlogImage1 from "../../Images/post-thumbnail-1.jpg";
+import BlogImage2 from "../../Images/post-thumbnail-2.jpg";
+import BlogImage3 from "../../Images/post-thumbnail-3.jpg";
+
+const Images = {
+  OrganicLogo,
+  Banner,
+  CategoryImage1,
+  CategoryImage2,
+  CategoryImage3,
+  CategoryImage4,
+  CategoryImage5,
+  CategoryImage6,
+  CategoryImage7,
+  CategoryImage8,
+  ProductThumb1,
+  ProductThumb2,
+  ProductThumb3,
+  ProductThumb4,
+  ProductThumb5,
+  ProductThumb6,
+  ProductThumb7,
+  ProductThumb8,
+  ProductThumb9,
+  ProductThumb10,
+  ProductThumb11,
+  ProductThumb12,
+  ProductThumb13,
+  ProductThumb14,
+  ProductThumb15,
+  ProductThumb16,
+  ProductThumb17,
+  ProductThumb18,
+  ProductThumb19,
+  ProductThumb20,
+  ProductThumb21,
+  ProductThumb22,
+  BgDiscountForm,
+  BlogImage1,
+  BlogImage2,
+  BlogImage3,
+};
+export default Images;
